pyproject.toml
[project]
name = "streamline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.35.0",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "streamlit-authenticator>=0.3.0",
    "bcrypt>=4.1.0",
]
